Add files via upload
Uploaded a compiled EPIC SNP manifest for fingerprinting using all 59 SNPs if imputed genotypes are available. Affy system is used to code corresponding methylation scores (ie if Methyl(beta) = 1, then Affy score is 0 (corresponding to ref allele) or 2 (corresponding to alternate allele) for a given snp). 

Data is compiled from an in-house project where DNA methylation (EPIC) and Genotyping (affymatrix) data are available for a cohort of 1,500 individuals. Part of this table is pulled from ttriche/infiniumSnps and credit must be given to Tim Triche.